+++ R/lmtp-package.R
@@ -1,4 +1,6 @@
-#' @importFrom stats runif as.formula coef glm plogis predict qlogis qnorm pnorm sd quantile var binomial gaussian na.omit weighted.mean
+#' @importFrom stats setNames ave runif as.formula coef glm plogis predict qlogis qnorm pnorm sd quantile var binomial gaussian na.omit weighted.mean
+#' @importFrom checkmate assert_character assert_list assert_number assert_function assert_subset assert_numeric assert_logical
+#' @importFrom origami make_folds
 #' @keywords internal
 "_PACKAGE"