+++ R/lmtp_control.R
@@ -12,9 +12,11 @@
 #' @param .learners_trt_folds \[\code{integer(1)}\]\cr
 #'  The number of cross-validation folds for \code{learners_trt}.
 #' @param .return_full_fits \[\code{logical(1)}\]\cr
-#'  Return full SuperLearner fits? Default is \code{FALSE}, return only SuperLearner weights.
-#' @param .B description
-#' @param .boot_seed description
+#'  Return full 'mlr3superlearner' fits? Default is \code{FALSE}, return only 'mlr3superlearner' weights.
+#' @param .discrete \[\code{logical(1)}\]\cr
+#'  Use discrete or ensemble super learner?
+#' @param .info \[\code{logical(1)}\]\cr
+#'  Print super learner fitting info to the console?
 #'
 #' @return A list of parameters controlling the estimation procedure.
 #' @export
@@ -26,13 +28,22 @@ lmtp_control <- function(.bound = 1e5,
                          .learners_outcome_folds = 10,
                          .learners_trt_folds = 10,
                          .return_full_fits = FALSE,
-                         .B = 1000,
-                         .boot_seed = NULL) {
+                         .discrete = TRUE,
+                         .info = FALSE) {
+
+  assert_number(.learners_outcome_folds, null.ok = TRUE)
+  assert_number(.learners_trt_folds, null.ok = TRUE)
+  assert_number(.bound)
+  assert_number(.trim, upper = 1)
+  assert_logical(.return_full_fits, len = 1)
+  assert_logical(.discrete, len = 1)
+  assert_logical(.info, len = 1)
+
   list(.bound = .bound,
        .trim = .trim,
        .learners_outcome_folds = .learners_outcome_folds,
        .learners_trt_folds = .learners_trt_folds,
        .return_full_fits = .return_full_fits,
-       .B = .B,
-       .boot_seed = .boot_seed)
+       .discrete = .discrete,
+       .info = .info)
 }